# domain-product-pricing-sdk/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:18080
  description: Generated server url
tags:
- name: Waivers
  description: Pricing waivers / promotional discounts stored as product configurations.
- name: Fees
  description: CQ queries and registration for product fees
- name: Pricing
  description: CQ queries and registration for product pricing
- name: Eligibility
  description: CQ queries and registration for product eligibility
paths:
  /api/v1/pricing/{pricingId}:
    put:
      tags:
      - Pricing
      summary: Amend pricing
      description: Amend rates/margins/tiers by creating a new effective version.
      operationId: amendPricing
      parameters:
      - name: pricingId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductPricingCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/pricing/waivers/{waiverId}:
    put:
      tags:
      - Waivers
      summary: Update waiver
      description: Replaces the spec of an existing waiver. The path productId ensures
        cross-tenant isolation.
      operationId: updateWaiver
      parameters:
      - name: waiverId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaiverSpec"
        required: true
      responses:
        "200":
          description: Updated.
          content:
            '*/*':
              schema:
                type: string
                format: uuid
    delete:
      tags:
      - Waivers
      summary: Remove waiver
      description: Deletes a waiver configuration.
      operationId: removeWaiver
      parameters:
      - name: waiverId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Deleted.
  /api/v1/pricing/fees/schemes/{schemeId}/components/{componentId}:
    put:
      tags:
      - Fees
      summary: Update fee rule
      description: Update a specific fee calculation rule.
      operationId: updateFeeRule
      parameters:
      - name: schemeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: componentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeeApplicationRuleCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/pricing:
    post:
      tags:
      - Pricing
      summary: Register pricing
      description: Create rates with tiers and effectiveFrom date.
      operationId: registerPricing
      parameters:
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterProductPricingCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/pricing/waivers:
    post:
      tags:
      - Waivers
      summary: Create waiver
      description: Attaches a pricing waiver to a product.
      operationId: createWaiver
      parameters:
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWaiverCommand"
        required: true
      responses:
        "200":
          description: Created; waiver id returned in the response body.
          content:
            '*/*':
              schema:
                type: string
                format: uuid
  /api/v1/pricing/schemes/{id}/simulate-impact:
    post:
      tags:
      - Pricing
      summary: Simulate scheme impact
      description: Projects portfolio-wide impact of a proposed pricing scheme change.
      operationId: simulateSchemeImpact
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateSchemeImpactRequest"
        required: true
      responses:
        "200":
          description: Per-product deltas and aggregate statistics.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SchemeImpactReport"
  /api/v1/pricing/fees/schemes:
    post:
      tags:
      - Fees
      summary: Define fee scheme
      description: Define fee types and calculation rules.
      operationId: defineFeeScheme
      parameters:
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterFeeSchemaCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/pricing/eligibility:
    post:
      tags:
      - Eligibility
      summary: Publish eligibility criteria
      description: "Publish eligibility rules (KYC/KYB, score, income, activity)."
      operationId: publishEligibility
      parameters:
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishEligibilityCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/pricing/eligibility/{eligibilityId}/evaluate:
    post:
      tags:
      - Eligibility
      summary: Evaluate eligibility
      description: Evaluate applicant facts and return fit/not-fit with reasons.
      operationId: evaluateEligibility
      parameters:
      - name: eligibilityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateEligibilityCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/pricing/eligibility/{eligibilityId}:
    patch:
      tags:
      - Eligibility
      summary: Adjust eligibility criteria
      description: Adjust criteria with versioning.
      operationId: adjustEligibility
      parameters:
      - name: eligibilityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustEligibilityCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/pricing/waivers/search:
    get:
      tags:
      - Waivers
      summary: Search waivers
      description: "Searches waivers for a product, optionally filtering by code substring\
        \ and active-on date."
      operationId: searchWaivers
      parameters:
      - name: productId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: codeContains
        in: query
        required: false
        schema:
          type: string
      - name: activeOn
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Matching waiver summaries.
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WaiverSummary"
  /api/v1/pricing/history/{entityId}:
    get:
      tags:
      - Pricing
      summary: Get pricing history
      description: "Best-effort audit timeline for a pricing entity (waiver, scheme)."
      operationId: getPricingHistory
      parameters:
      - name: entityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "Timeline entries, oldest first."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PricingHistoryResponse"
components:
  schemas:
    UpdateProductPricingCommand:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
        productConfigurationId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        configType:
          type: string
          enum:
          - PRICING
          - LIMITS
          - FEATURES
          - CUSTOM
        configKey:
          type: string
        configValue:
          type: string
        productPricingId:
          type: string
          format: uuid
    WaiverSpec:
      required:
      - code
      type: object
      properties:
        code:
          maxLength: 40
          minLength: 0
          type: string
          description: Short waiver code (alphanumeric + underscore). Used as the
            config key suffix.
        description:
          maxLength: 500
          minLength: 0
          type: string
          description: Human-readable description.
        eligibilityFilter:
          type: string
          description: Optional eligibility filter expression (JSON blob interpreted
            by pricing engine).
        adjustment:
          type: number
          description: Adjustment applied when the waiver is triggered (e.g. -0.005
            for -50 bps).
        validFrom:
          type: string
          description: Start date (inclusive).
          format: date
        validTo:
          type: string
          description: End date (inclusive).
          format: date
      description: Full specification for a pricing waiver.
    UpdateFeeApplicationRuleCommand:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
        productConfigurationId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        configType:
          type: string
          enum:
          - PRICING
          - LIMITS
          - FEATURES
          - CUSTOM
        configKey:
          type: string
        configValue:
          type: string
        feeApplicationRuleId:
          type: string
          format: uuid
        feeComponentId:
          type: string
          format: uuid
    RegisterProductPricingCommand:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
        productConfigurationId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        configType:
          type: string
          enum:
          - PRICING
          - LIMITS
          - FEATURES
          - CUSTOM
        configKey:
          type: string
        configValue:
          type: string
    CreateWaiverCommand:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        spec:
          $ref: "#/components/schemas/WaiverSpec"
        tenantId:
          type: string
          format: uuid
    PricingSchemeUpdate:
      type: object
      properties:
        newInterestRate:
          type: number
          description: "New nominal annual interest rate (decimal, e.g. 0.045)."
        newFixedFee:
          type: number
          description: New fixed fee applied per transaction.
        newEffectiveApr:
          type: number
          description: New effective APR disclosed to customers.
      description: Proposed pricing scheme change fed into the impact simulator.
    SimulateSchemeImpactRequest:
      required:
      - proposedChange
      type: object
      properties:
        proposedChange:
          $ref: "#/components/schemas/PricingSchemeUpdate"
      description: "Impact-simulation request: proposed change to a pricing scheme."
    ProductImpact:
      type: object
      properties:
        productId:
          type: string
          description: Product identifier.
          format: uuid
        currentMonthlyPayment:
          type: number
          description: Current monthly payment under the existing scheme (reference
            scenario).
        newMonthlyPayment:
          type: number
          description: Projected monthly payment under the proposed scheme.
        deltaPercent:
          type: number
          description: "Relative delta, e.g. 0.05 means +5%."
      description: Impact of a pricing scheme change on a single product.
    SchemeImpactReport:
      type: object
      properties:
        schemeId:
          type: string
          description: Scheme id under simulation.
          format: uuid
        affectedProductCount:
          type: integer
          description: Number of products touched by the change.
          format: int32
        productImpacts:
          type: array
          description: Per-product impact breakdown.
          items:
            $ref: "#/components/schemas/ProductImpact"
        stats:
          $ref: "#/components/schemas/SchemeImpactStats"
      description: Projected impact of a pricing scheme change across the product
        portfolio.
    SchemeImpactStats:
      type: object
      properties:
        avgDeltaPercent:
          type: number
          description: Average delta across affected products.
        maxIncreasePercent:
          type: number
          description: Largest positive delta observed.
        maxDecreasePercent:
          type: number
          description: Largest negative delta observed (as a negative number).
      description: Aggregate statistics across all affected products.
    RegisterFeeApplicationRuleCommand:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
        productConfigurationId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        configType:
          type: string
          enum:
          - PRICING
          - LIMITS
          - FEATURES
          - CUSTOM
        configKey:
          type: string
        configValue:
          type: string
        feeComponentId:
          type: string
          format: uuid
    RegisterFeeComponentCommand:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
        productConfigurationId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        configType:
          type: string
          enum:
          - PRICING
          - LIMITS
          - FEATURES
          - CUSTOM
        configKey:
          type: string
        configValue:
          type: string
        feeStructureId:
          type: string
          format: uuid
    RegisterFeeSchemaCommand:
      type: object
      properties:
        feeApplicationRule:
          $ref: "#/components/schemas/RegisterFeeApplicationRuleCommand"
        feeStructure:
          $ref: "#/components/schemas/RegisterFeeStructureCommand"
        feeComponent:
          $ref: "#/components/schemas/RegisterFeeComponentCommand"
        productFeeStructure:
          $ref: "#/components/schemas/RegisterProductFeeStructureCommand"
    RegisterFeeStructureCommand:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
        productConfigurationId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        configType:
          type: string
          enum:
          - PRICING
          - LIMITS
          - FEATURES
          - CUSTOM
        configKey:
          type: string
        configValue:
          type: string
    RegisterProductFeeStructureCommand:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
        productConfigurationId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        configType:
          type: string
          enum:
          - PRICING
          - LIMITS
          - FEATURES
          - CUSTOM
        configKey:
          type: string
        configValue:
          type: string
        feeStructureId:
          type: string
          format: uuid
    PublishEligibilityCommand:
      type: object
    EvaluateEligibilityCommand:
      type: object
    AdjustEligibilityCommand:
      type: object
    WaiverSummary:
      type: object
      properties:
        waiverId:
          type: string
          description: Waiver (configuration) identifier.
          format: uuid
        productId:
          type: string
          description: Product the waiver applies to.
          format: uuid
        spec:
          $ref: "#/components/schemas/WaiverSpec"
      description: Summary of a pricing waiver as returned from a search.
    HistoryEntry:
      type: object
      properties:
        timestamp:
          type: string
          description: When the action happened.
          format: date-time
        action:
          type: string
          description: "Action taken: CREATED, UPDATED or DELETED."
        actor:
          type: string
          description: "Actor identifier. TODO: wire to centralized audit store."
        snapshot:
          type: string
          description: Serialized snapshot of the entity at this point in time.
      description: Single entry in a pricing entity's audit timeline.
    PricingHistoryResponse:
      type: object
      properties:
        entityId:
          type: string
          description: Pricing entity identifier (configuration id).
          format: uuid
        entityType:
          type: string
          description: "Entity type inferred from the configuration, e.g. PRICING_WAIVER."
        entries:
          type: array
          description: Chronological audit entries.
          items:
            $ref: "#/components/schemas/HistoryEntry"
      description: Audit timeline for a pricing entity.